Fix bug where renderable in region is destroyed.
When a renderable is destroyed that is added via `renderRegion`, the region still has reference to the element. This element has been detached from the DOM, so a future call to render the region with a new renderable fails, because we can't swap.

Instead, listen to the `beforeDestroy`hook and swap the renderable to be destroyed element with an outlet el, so that we still have a reference to an element in the DOM.